address review: scope the owner check, the guidance, the generation, and the retire seam
**A carried prompt leaked into a new chat.** The ownership check required both
ids to be non-null, so a new chat, whose `activeConversationId` is still null,
did not count as a mismatch: with an empty transcript the old card docked above
the new chat's composer, wired to whichever assistant had become active. A
missing active conversation is now a mismatch rather than an absence of
opinion.

**The model was told about a card that was never raised.** The generation guard
covered the event, the registry and the persisted marker, but `notifyParent`
still appended `ACP_AUTH_RECOVERY_GUIDANCE` on the anchor alone. It now turns
on the same predicate.

**The generation was sampled too late.** Recording it at session registration
missed a replacement that landed between `prepareAgentEnv` reading the token
and the caller reaching `manager.spawn`, leaving the session claiming a
generation newer than the token it actually held. It now travels on the config
`prepareAgentEnv` returns, stamped with the credential read.

**Only one of three writers retired the markers.** `assistant credentials set`
and the headless secure prompt reach the same vault field without going
through `storeAcpClaudeToken`, so a token repaired either way left the markers
and the registry standing. The behaviour moves to one `retireAcpAuthRecovery`
that all three call, which is also where the generation counter now lives.
